move row-render.ts to shared.ts
