+++ packages/web/src/shared.ts
